+++ libs/nestjs/authentication/src/interfaces/authentication-module.interface.ts
@@ -1,8 +1,10 @@
 import { JwtModuleOptions } from '@nestjs/jwt';
 import { IAuthModuleOptions } from '@nestjs/passport';
-import { User } from '@prisma/client';
+import { CookieOptions } from 'express';
 import { StrategyOptions } from 'passport-jwt';
 import { IStrategyOptionsWithRequest } from 'passport-local';
+
+import { User } from '../../prisma/client';
 
 export interface AuthenticationOptions {
   login: {
@@ -11,6 +13,7 @@ export interface AuthenticationOptions {
   cookies: {
     cookieName: string;
     queryParamName: string;
+    options: CookieOptions;
   };
   strategy: {
     jwt: StrategyOptions;