+++ libs/nestjs/authentication/test/integration/authentication.module.spec.ts
@@ -4,22 +4,25 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { mockDeep, MockProxy } from 'jest-mock-extended';
 import * as request from 'supertest';
 
-import {
-  AuthenticationModule,
-  AuthenticationService,
-  JwtGlobalAuthGuard,
-} from '../../src';
-import { AUTHENTICATION_OPTIONS } from '../../src/config';
-import { AuthenticationEndpointMockController } from '../mock/authentication-endpoint-mock.controller';
-
-import { mockUserFactory } from '@generated/models/mock';
+import { mockUserFactory } from '../../generated/models/mock';
 import {
   USER_DATABASE_SERVICE,
   USER_SERVICE,
   UserDatabaseService,
   UserService,
-} from '@generated/nestjs-models-common';
-import { mockUserServiceFactory } from '@generated/nestjs-models-common/mock';
+} from '../../generated/nestjs-models-common';
+import { mockUserServiceFactory } from '../../generated/nestjs-models-common/mock';
+import {
+  AuthenticationModule,
+  AuthenticationService,
+  JwtGlobalAuthGuard,
+} from '../../src';
+import {
+  AUTHENTICATION_OPTIONS,
+  AUTHENTICATION_QUERY_PARAM_NAME,
+} from '../../src/config';
+import { AuthenticationEndpointMockController } from '../mock/authentication-endpoint-mock.controller';
+
 import { LoggerModule } from '@tractr/nestjs-core';
 
 describe('Authentication Module (integration)', () => {
@@ -89,7 +92,7 @@ describe('Authentication Module (integration)', () => {
       mockUserService.findUnique.mockResolvedValue({
         ...mockUser,
         password: hashPassword,
-      });
+      } as never);
 
       const response = await request(app.getHttpServer())
         .post('/login')
@@ -107,11 +110,27 @@ describe('Authentication Module (integration)', () => {
       const mockUser = mockUserFactory();
       const accessToken = await authenticationService.createUserJWT(mockUser);
 
-      mockUserService.findUnique.mockResolvedValue(mockUser);
+      mockUserService.findUnique.mockResolvedValue(mockUser as never);
 
       const response = await request(app.getHttpServer())
         .get('/me')
         .set('Authorization', `bearer ${accessToken}`);
+
+      expect(response.status).toBe(200);
+      expect(response.body).toEqual(JSON.parse(JSON.stringify(mockUser)));
+    });
+    it('/me get the user information back and use the query param auth strategy', async () => {
+      const authenticationService = app.get<AuthenticationService>(
+        AuthenticationService,
+      );
+      const mockUser = mockUserFactory();
+      const accessToken = await authenticationService.createUserJWT(mockUser);
+
+      mockUserService.findUnique.mockResolvedValue(mockUser as never);
+
+      const response = await request(app.getHttpServer()).get(
+        `/me?${AUTHENTICATION_QUERY_PARAM_NAME}=${accessToken}`,
+      );
 
       expect(response.status).toBe(200);
       expect(response.body).toEqual(JSON.parse(JSON.stringify(mockUser)));