+++ libs/nestjs/core/src/helpers/index.ts
@@ -1,5 +1,6 @@
 export * from './is-class.helper';
 export * from './is-development.helper';
+export * from './is-production.helper';
 export * from './module-metadata.helper';
 export * from './module-options.helper';
 export * from './module-override.helper';