+++ libs/nestjs/mailer/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/mailer.module';
+export * from './lib/interfaces';
+export * from './lib/services';