+++ apps/api/CHANGELOG.md
@@ -1,7 +1,37 @@
 # Changelog
 
-This file was generated using
-[@jscutlery/semver](https://github.com/jscutlery/semver).
+This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
+
+# [1.14.0](https://github.com/tractr/stack/compare/api-1.13.0...api-1.14.0) (2021-07-05)
+
+
+### Features
+
+* create nestjs mailer module ([#147](https://github.com/tractr/stack/issues/147)) ([40ef77f](https://github.com/tractr/stack/commit/40ef77f85f9e5785c918633ac2ba2f223d990f87))
+
+
+
+# [1.13.0](https://github.com/tractr/stack/compare/api-1.12.0...api-1.13.0) (2021-06-30)
+
+
+### Bug Fixes
+
+* update dockerfile to be built and deploy by the ci ([d71f395](https://github.com/tractr/stack/commit/d71f395ae0fc21506f34ea8313ee1fecf63dd4fc))
+
+
+### Features
+
+* add apps version and release back to get docker tags ([84332a5](https://github.com/tractr/stack/commit/84332a5fa01a2352d30a7d3674f1916314483c3c))
+
+
+
+# [1.12.0](https://github.com/tractr/stack/compare/api-1.11.0...api-1.12.0) (2021-06-24)
+
+### Features
+
+- change generated library configurations
+  ([#146](https://github.com/tractr/stack/issues/146))
+  ([3ee951e](https://github.com/tractr/stack/commit/3ee951e998b6e7d4bcbf9f66d4216ce155958cea))
 
 # [1.11.0](https://github.com/tractr/stack/compare/api-1.10.2...api-1.11.0) (2021-06-16)