+++ libs/hapify/templates/nestjs-models-common/CHANGELOG.md
@@ -3,6 +3,14 @@
 This file was generated using
 [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.4.0](https://github.com/tractr/stack/compare/hapify-templates-nestjs-models-common-1.3.0...hapify-templates-nestjs-models-common-1.4.0) (2021-06-24)
+
+### Features
+
+- change generated library configurations
+  ([#146](https://github.com/tractr/stack/issues/146))
+  ([3ee951e](https://github.com/tractr/stack/commit/3ee951e998b6e7d4bcbf9f66d4216ce155958cea))
+
 # [1.3.0](https://github.com/tractr/stack/compare/hapify-templates-nestjs-models-common-1.2.6...hapify-templates-nestjs-models-common-1.3.0) (2021-06-16)
 
 ### Features